+++ tests/test_models/test_mutables/test_mutable_channel/test_units/test_mutable_channel_units.py
@@ -11,8 +11,8 @@ from mmrazor.models.mutables.mutable_channel import (
 from mmrazor.models.mutables.mutable_channel.units.channel_unit import (  # noqa
     Channel, ChannelUnit)
 from mmrazor.structures.graph import ModuleGraph as ModuleGraph
-from ....data.models import LineModel
-from ....test_core.test_graph.test_graph import TestGraph
+from .....data.models import LineModel
+from .....test_core.test_graph.test_graph import TestGraph
 
 MUTABLE_CFG = dict(type='SimpleMutablechannel')
 PARSE_CFG = dict(
